[core-rest-pipeline] Merged bearerTokenChallengeAuthenticationPolicy into bearerTokenAuthenticationPolicy
